task21/requirement.txt
absl-py==0.7.1
appdirs==1.4.4
arrow==0.15.5
asgiref==3.2.5
asn1crypto==1.0.1
astor==0.8.0
astroid==2.3.3
attrs==19.1.0
awsebcli==3.19.2
backcall==0.1.0
beautifulsoup4==4.9.2
bleach==3.1.0
blis==0.4.1
botocore==1.19.39
Bottleneck==1.3.1
cachetools==4.0.0
captcha==0.3
catalogue==0.0.8
cement==2.8.2
certifi==2019.11.28
cffi==1.12.3
chardet==3.0.4
ChatterBot==1.0.8
cloudpickle==1.2.2
colorama==0.4.3
comtypes==1.1.7
cryptography==2.7
cvlib==0.2.2
cvxopt==1.2.3
cycler==0.10.0
cymem==2.0.3
Cython==0.29.14
decorator==4.4.0
defusedxml==0.6.0
dill==0.3.1.1
distlib==0.3.1
dj-database-url==0.5.0
Django==3.0.7
django-allauth==0.41.0
django-crispy-forms==1.9.0
django-extensions==3.0.9
django-heroku==0.3.1
django-markitup==4.0.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.11
djangorestframework==3.12.2
dlx==1.0.4
docloud==1.0.375
docopt==0.6.2
docplex==2.10.155
entrypoints==0.3
facenet==1.0.5
fastai==1.0.59
fastdtw==0.3.4
fastprogress==0.1.22
filelock==3.0.12
flirimageextractor==1.3.0
future==0.16.0
gast==0.2.2
google-auth==1.10.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.7
grpcio==1.22.0
gunicorn==20.0.4
h5py==2.9.0
idna==2.9
imageai==2.1.5
imageio==2.6.1
importlib-metadata==1.3.0
imutils==0.5.3
inflection==0.3.1
intel-openmp==2019.0
ipykernel==5.1.1
ipython==7.6.1
ipython-genutils==0.2.0
ipywidgets==7.5.0
isort==4.3.21
jedi==0.14.1
Jinja2==2.10.1
jmespath==0.10.0
joblib==0.13.2
Js2Py==0.70
json5==0.8.5
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
jupyterlab==1.0.4
jupyterlab-server==1.0.0
Keras==2.0.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
Markdown==3.1.1
MarkupSafe==1.1.1
marshmallow==3.5.0
marshmallow-polyfield==5.8
mathparse==0.1.2
matplotlib==3.1.1
mccabe==0.6.1
mistune==0.8.4
mkl==2019.0
mock==3.0.5
more-itertools==5.0.0
mpmath==1.1.0
mtcnn==0.0.9
murmurhash==1.0.2
mysqlclient==2.0.1
nbconvert==5.5.0
nbformat==4.4.0
ndg-httpsclient==0.5.1
nest-asyncio==1.0.0
networkx==2.3
nltk==3.4.5
notebook==5.7.8
ntlm-auth==1.4.0
numexpr==2.7.0
numpy==1.16.4
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
object-detection==0.1
opencv-python==4.1.0.25
opt-einsum==3.1.0
packaging==19.2
pandas==0.25.0
pandocfilters==1.4.2
parso==0.5.1
pathspec==0.5.9
pickleshare==0.7.5
Pillow==7.0.0
pip-utils==0.0.6
pipwin==0.5.0
plac==1.1.3
ply==3.11
preshed==3.0.2
prince==0.6.3
progressbar==2.5
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.9.0
psutil==5.6.3
psycopg2==2.8.6
psycopg2-binary==2.8.2
pyasn1==0.4.7
pyasn1-modules==0.2.7
PyAudio @ file:///C:/Users/gaurav/pipwin/PyAudio-0.2.11-cp37-cp37m-win_amd64.whl
pybind11==2.4.3
pycocotools==2.0
pycparser==2.19
pyflakes==2.1.1
Pygments==2.4.2
pyjsparser==2.7.1
pylint==2.4.4
pyOpenSSL==19.0.0
pyparsing==2.4.0
pypiwin32==223
PyPrind==2.11.2
pyrsistent==0.15.3
pySmartDL==1.3.4
PySocks==1.7.1
python-dateutil==2.8.0
python3-openid==3.1.0
pyttsx3==2.90
pytz==2019.3
pyvim==3.0.2
PyWavelets==1.1.1
pywin32==228
pywinpty==0.5.5
PyYAML==5.3.1
pyzmq==18.0.2
qiskit==0.15.0
qiskit-aer==0.4.0
qiskit-aqua==0.6.4
qiskit-ibmq-provider==0.4.6
qiskit-ignis==0.2.0
qiskit-terra==0.12.0
qtconsole==4.5.1
Quandl==3.4.8
requests==2.23.0
requests-ntlm==1.1.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-image==0.16.2
scikit-learn==0.21.3
scipy==1.1.0
semantic-version==2.5.0
Send2Trash==1.5.0
six==1.14.0
sklearn==0.0
soupsieve==1.9.5
spacy==2.2.3
SpeechRecognition==3.8.1
SQLAlchemy==1.3.19
sqlparse==0.3.1
srsly==0.2.0
stripe==2.43.0
sympy==1.4
tensorboard==1.15.0
tensorflow==1.13.1
tensorflow-cpu==1.15.0
tensorflow-cpu-estimator==1.15.1
tensorflow-estimator==1.13.0
tensorflow-probability==0.6.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
Theano==1.0.4
thinc==7.3.1
torch==1.0.1
torchvision==0.1.8
tornado==6.0.3
tqdm==4.40.2
traitlets==4.3.2
typed-ast==1.4.1
tzlocal==2.1
urllib3==1.25.6
utils==1.0.0
virtualenv==20.1.0
wasabi==0.4.2
wcwidth==0.1.7
web.py==0.40.dev1
webencodings==0.5.1
websockets==7.0
Werkzeug==0.15.4
whitenoise==5.2.0
widgetsnbextension==3.5.0
wrapt==1.11.2
zipp==0.6.0
